+++ core/design-system/src/main/java/com/moneymong/moneymong/design_system/component/textfield/visualtransformation/PriceVisualTransformation.kt
@@ -1,9 +1,10 @@
-package com.moneymong.moneymong.design_system.component.textfield.price
+package com.moneymong.moneymong.design_system.component.textfield.visualtransformation
 
 import androidx.compose.ui.text.AnnotatedString
 import androidx.compose.ui.text.input.OffsetMapping
 import androidx.compose.ui.text.input.TransformedText
 import androidx.compose.ui.text.input.VisualTransformation
+import com.moneymong.moneymong.design_system.component.textfield.util.PriceType
 
 class PriceVisualTransformation(private val type: PriceType) : VisualTransformation {
     override fun filter(text: AnnotatedString): TransformedText {
